Add the MIT license since that is what most dependencies are using
I am not entirely clear on the legality of licenses, but as I am not
making any sort of profit or exchanging any money for people to use this
product, I do not think there will be any issues. If there are, I am
writing this commit description to demonstrate that I do not intend to
infringe on any sort of intellectual property laws or any licenses, and
I will happily cooperate with any offended parties to resolve these
issues.